refactor: image folder component and image
