+++ apps/web/app/components/base-component/index.ts
@@ -1,5 +1,2 @@
 export * from "./header";
 export * from "./sidebar";
-export * from "./display-file";
-export * from "./folder";
-export * from "./create-folder-button";
