+++ apps/web/app/components/image-folder/delete-folder-button.tsx
