+++ apps/web/app/components/image-folder/display-file.tsx
