+++ apps/web/app/components/image-folder/index.ts
@@ -1,0 +1,5 @@
+export * from "./create-folder-button";
+export * from "./delete-folder-button";
+export * from "./display-file";
+export * from "./folder";
+export * from "./rename-folder-button";
