+++ apps/web/app/components/image-folder/rename-folder-button.tsx
